Move `EnemyPokemon` to `src/field/enemy-pokemon.ts`
