+++ src/data/mystery-encounters/encounters/berries-abound-encounter.ts
@@ -10,7 +10,7 @@ import {
 } from "#app/data/mystery-encounters/utils/encounter-phase-utils";
 import type { PlayerPokemon } from "#app/field/pokemon";
 import type { Pokemon } from "#app/field/pokemon";
-import { EnemyPokemon } from "#app/field/pokemon";
+import { EnemyPokemon } from "#app/field/enemy-pokemon";
 import type { BerryModifierType, ModifierTypeOption } from "#app/modifier/modifier-type";
 import { getPartyLuckValue, regenerateModifierPoolThresholds } from "#app/modifier/modifier-type";
 import { modifierTypes } from "#app/modifier/modifier-types";
