+++ src/phases/attempt-capture-phase.ts
@@ -8,7 +8,7 @@ import {
   getPokeballTintColor,
 } from "#app/data/pokeball";
 import { getStatusEffectCatchRateMultiplier } from "#app/data/status-effect";
-import { type EnemyPokemon } from "#app/field/pokemon";
+import { type EnemyPokemon } from "#app/field/enemy-pokemon";
 import { globalScene } from "#app/global-scene";
 import { getPokemonNameWithAffix } from "#app/messages";
 import { PokemonPhase } from "#app/phases/abstract-pokemon-phase";
