+++ src/system/pokemon-data.ts
@@ -5,7 +5,8 @@ import type { Nature } from "#enums/nature";
 import type { PokeballType } from "#enums/pokeball-type";
 import { getPokemonSpecies } from "#app/utils/pokemon-species-utils";
 import { Status } from "../data/status-effect";
-import { type Pokemon, EnemyPokemon } from "#app/field/pokemon";
+import { type Pokemon } from "#app/field/pokemon";
+import { EnemyPokemon } from "#app/field/enemy-pokemon";
 import { PokemonSummonData } from "#app/field/pokemon-summon-data";
 import { PokemonMove } from "#app/field/pokemon-move";
 import { TrainerSlot } from "#enums/trainer-slot";
