+++ test/abilities/poison-touch.test.ts
@@ -1,6 +1,6 @@
 import { BattlerIndex } from "#enums/battler-index";
 import { type PostAttackApplyStatusEffectAbAttr } from "#app/data/abilities/ab-attrs/post-attack-apply-status-effect-ab-attr";
-import type { EnemyPokemon } from "#app/field/pokemon";
+import type { EnemyPokemon } from "#app/field/enemy-pokemon";
 import { globalScene } from "#app/global-scene";
 import { AbilityId } from "#enums/ability-id";
 import { MoveId } from "#enums/move-id";
