+++ test/ai/enemy-command.test.ts
@@ -4,7 +4,7 @@ import { MoveCategory } from "#enums/move-category";
 import { AbilityId } from "#enums/ability-id";
 import { MoveId } from "#enums/move-id";
 import { SpeciesId } from "#enums/species-id";
-import type { EnemyPokemon } from "#app/field/pokemon";
+import type { EnemyPokemon } from "#app/field/enemy-pokemon";
 import { AiType } from "#enums/ai-type";
 import { randSeedInt } from "#app/utils";
 import { GameManager } from "#test/test-utils/gameManager";
